Added link system to allow components addition
On NavBar component with buttonList variable on Line 33, add the name you want the page to be to the array (this will be refered to as PATH).

On index.tsx add the following Tag
<Route path ="/PATH" element = {<YourComponent/>}/>

around line 51 - 53 in between the <Routes></Routes>

When your PATH appears on the navbar clicking on it should make your component appear.